requirements.txt
streamlit==1.30.0
sentence-transformers>=5.2.0
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
